[network-setting] refactor : 트위치 관련 json 데이터 트위치 패키지로 이동
