+++ data/src/main/java/com/hegunhee/data/data/json/twitch/Pagination.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.data.data.json
+package com.hegunhee.data.data.json.twitch
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
