+++ data/src/main/java/com/hegunhee/data/mapper/SearchApiDataMapper.kt
@@ -1,6 +1,6 @@
 package com.hegunhee.data.mapper
 
-import com.hegunhee.data.data.json.SearchApiData
+import com.hegunhee.data.data.json.twitch.SearchApiData
 import com.hegunhee.domain.model.SearchData
