+++ data/src/main/java/com/hegunhee/data/network/TwitchAuthService.kt
@@ -1,7 +1,7 @@
 package com.hegunhee.data.network
 
 import com.hegunhee.data.BuildConfig
-import com.hegunhee.data.data.json.AuthToken
+import com.hegunhee.data.data.json.twitch.TwitchAuthToken
 import retrofit2.http.POST
 import retrofit2.http.Query
 
@@ -12,6 +12,6 @@ interface TwitchAuthService {
         @Query("client_id") clientId : String = BuildConfig.clientId,
         @Query("client_secret") clientSecret : String = BuildConfig.clientSecret,
         @Query("grant_type") grantType : String = "client_credentials"
-    ) : AuthToken
+    ) : TwitchAuthToken
 
 }
